Updated schema testing code
One thing that's nice about this new system is that we can have a
general test that checks if data that was present in the first version
is still present in the current version.  That means we don't need to
define structs to test each different version and the individual upgrade
tests can focus on the specifics for that upgrade.