Make `cargo uniffi-bindgen` alias work when executed from a crate.
Previously, our special `cargo uniffi-bindgen` alias would work correctly
when executed from the workspace root directory, but would not work when
executed in the root directory of a crate (because from that location,
`cargo run` tries to run a binary from the crate you're in rather than
from the containing workspace).

This tweaks the alias to explicitly specify the package containing
the command, so it should work from any directory.