Include fxa-client and logins in the XCFramework build.
This is mostly to see if we can successfully include things that depend
on NSS and SQLCipher. It builds and links OK but needs to be tested on
various hardware and simulators.

I've removed redundant copies of `.h` files, opting to instead copy
them from their canonical home or generate them on demand.