Prep work for login struct refactor
This is the first step to try to refactor the login structs based on
trying to get things to work with android-components.  The comments at
the top of logins.rs describe where I want it to end up.  Nothing is
implemented yet, I'm trying to take it step by step.

- Reordered the structs in login.rs.  I tried to group them together a
  bit nicer.  In particular, all the ValidateAndFixup impls are
  together.
- Rename LoginEntry variables to "entry" rather than login.