I think this should fix the dependency check issue
The main issue is that we had a manual override for `humantime` in
PACKAGE_METADATA_FIXUPS, but the `humantime` package has been updated
and it doesn't match anymore.

I removed the PACKAGE_METADATA_FIXUPS entry.  I don't think we need it
anymore now since `humantime` now has a license file in it's repo.