Exit verify-ios-env immediately on error
If one of the called scripts fails, e.g. because some tooling is missing,
it would still print the "Looks good" message, which is confusing.
Now it just error-exits if one of the called scripts error-exits (which
btw prints an error message and if that's the last thing printed its
more noticable)