Move find_login_to_update() into LoginsDb
The original idea was that we would feed the output of
`get_by_base_domain()` into this function, but there are a couple
differences between the logins returned by that function and the logins
returned by `find_dupes_ignoring_username()` which it's supposed to
replace.  The simplest fix I could find was to have the function also
run the query and make the query match the old
`find_dupes_ignoring_username()`.