Moving code out of BookmarksEngine and HistoryEngine
- Moved the engine methods other than the `SyncEngine` ones to top-level
  functions.
- Made the `Merger` struct input a PlacesDb and SqlInterruptScope instead
  of a BookmarksEngine

I want to refactor the sync engine code to use an Arc<Mutex<PlacesDb>>,
similar to how the other sync engines work.  This means it would lock
the mutex around it's methods.  However, I couldn't get this to work
with the import code because it calls these methods and uses the Merger
struct, mixed in with its own queries.  I want the importer code to take
lock around that entire operation, but that would cause a deadlock if
the methods also locked the mutex.

Also updated the `SqlInterruptScope` code.  I believe the current code
is basically broken.  I think `PlacesApi` should be the struct to manage
things, not the sync connection which gets created/dropped as the app
runs.