Fix for fetch_login_data() (#4667)
Added a test that caused the wrong guid issue.  I'm pretty sure this is
what's causing the sentry issues -- at least some of them.

The issue was that we were calling `each_chunk_mapped()` with the
expectation that the `records` and `sync_data` vecs matched up
(`records[i].0.id == sync_data[i].guid`).  However, we only push items
to `sync_data` if `from_payload()` returns `Ok`.  If it returns an error
for any item then this fails.

I think this system could use a rework, but this should fix the error
for now.