Added features to support application-services branch builds
Added code that copies data from `local.properties` to the gradle
instance, like android-components does.

If `branchBuild.application-services.version` is set in
`local.properties`, then we set the app-services version to that exact
string.

Updated `substitute-local-appservices.gradle` to also support that
property (in this case it's set in the `android-components` directory).

The plan is to use this feature to build/test fenix using a particular
checkout of application-services, with the version set to the git commit
ID.