Implementing syncing interruption for real
Added the `interrupt_scope::InterruptScope` type which provides a
simplified API for interrupt handling.  One difference between this and
`SqlInterruptScope` is the counter is a static variable, this allows us
to create a new `InterruptScope` whenever we want, which simplied
refactoring some of the older code.

Updated the `SyncEngine` implemenations so each one stores an
`InterruptScope` to check.  Removed older/unused interruption code.

For `SyncManager` methods, where we create multiple sync engines, the
`SyncManager` creates one scope, then passes a clone to each
`SyncEngine`. This avoids any chance of some engines getting interrupted
and some not.

One thing that `InterruptScope` doesn't handle is SQL interruption.
Added places code to handle that since it's the only component that has
a dedicated sync connection.  At some point we could update the other
components to match and standardize this somehow, but for now it seems
okay to special-case places.

Made the code to use `interrupt_scope` as the field/argument name.  The existing
code was split between this one and `scope`.  I figured the extra descriptive
value was worth the extra chars.

There's still some `PlacesDb` code that uses the older `SqlInterruptScope` and
`SqlInterruptHandler` classes.  I left that alone because I think it's used for
interrupting autocomplete and I didn't want to get into that code at this
point.