Attempt at fixing the `database is locked` error (#4762)
The error was happening while opening the database so I added a
places-utils test to stress-test opening sync connections.  I was able
to reproduce the error and tracked down the source to the `INSERT OR
IGNORE` statements inside the `create_synced_bookmark_roots()` call.

The error happens when attempting to get a write lock for a transaction,
after the it already has a read lock.  This is causing SQL_BUSY errors
(which has a misleading "database is locked" error message).

This issue happened in two different ways:
  - When a separate thread held a write lock on the writer DB.  Note
    that the thread only held the lock for 2 seconds, which is short
    enough that other threads should just wait for the lock rather than
    timing out.

  - When multiple threads tried to open a sync connection at once.
    Because get_sync_connection uses a mutex this happened sequentially,
    but it still triggered the issue if open attempts happened
    immediately after the connection was dropped. Adding a 10ms sleep
    stopped the error.

Interestingly, if we used `BEGIN IMMEDIATE` to start the transaction,
which immediately tries to take the write lock, then the issue wouldn't
appear. The thread would just wait until the lock was available.  The
issue only happened when we took the read lock first, then tried to take
a write lock.

Fixed this by using the SQLITE_OPEN_SHARED_CACHE flag when opening the
connection.  This changes how sqlite handles locking and fixes the
problem.  It also seems good in general.  The sqlite site says it "can
significantly reduce the quantity of memory and IO required by the
system".

Also added some extra logging to the `open_database` code.