Attempt at fixing the `database is locked` error (#4762)
Added a places-utils test that opens sync connections in threads.  All
this needed to error out was:
  - The database was already created
  - 2 threads tried to open a sync connection. The Mutex in PlacesApi
    ensures that this happens sequentially, but if the second thread
    tried to immediately open the connection after the first one, then I
    would see an error.  Adding a 10ms delay in-between seemed to fix
    things.

After some investigation I narrowed it down to the `INSERT OR IGNORE`
statements inside the `create_synced_bookmark_roots()` call.  I think
the fact that that query is both a read and write causes some extra
locking to happen. I'm not sure exactly what's going on here.  Maybe the
disk needs to be flushed after the first connection closes the DB, but
`PRAGMA database.wal_checkpoint` didn't work.

In any case, it seems like we could create the roots when we are first
initializing the schema with the read/write connection so I just moved
the call there and it seemed to fix things for me.