Use BEGIN IMMEDIATE in `open_database`
This should fix #4762.  `PlacesDb` was already using immediate
transactions for the sync connection, so we just need to make sure the
database initialization code uses it to.

I think BEGIN IMMEDIATE makes sense in the general case too.  Usually
there should only be one connection accessing the database when it's
initialized, which means it won't really matter.  If we do have multiple
connections that are trying to access the database, I think it's worth
taking a small hit in concurrency in order to eliminate the possibility
of SQL_BUSY errors in cases like this.