Refactoring how `PlacesApi` stores it's databases
Refactored `PlacesApi` to store an Arc to it's database connections.
This is going to be useful when implementing interrupt handling.

This changes how we handle read-only and read-write connections.
Instead of being able to create unlimited read connections and one write
connection that's passed back and forth between the consumer code and
`PlacesApi`, there`just 1 read connection and 1 write connection and the
consumer can request an `Arc` to them.  I think this is okay because:

  - On both Android and iOS, the wrapper class creates a single
    connection of both types.  So we don't really need to support
    multiple read connections or ensure there aren't multiple write
    connections.
  - `PlacesConnection` wraps it's `PlacesDb` in a Mutex, so even the
    consumer code created multiple connections it would be fairly safe.

Added some testing code to make the existing tests continue to work.

Removed the tests that check that a reader/writer access the same
database.  It seems unnecessary to test that 2 sqlite connections point
to the same underlying DB.  Also, these were testing in-memory
databases, so it's not clear that they would catch an actual error.