Move Rust setup a toolchain task
Created a rust toolchain task that installs the rust components we use
for building and uploads them as artifacts.  This needs to run before
the other toolchain tasks, so I made a new kind called `early-toolchain`
for it.

Made the other tasks depend on this, and have them download and unpack
the artifact before starting.

This allows us to do this setup once rather than multiple times.

Moved functionality from `rustup-setup.sh` to other places:
  - The Dockerfile now sets the ENV variables and installs `rustup`.
    Note that we now always set `-Dwarnings`, before we didn't set if
    for releases because of the corner case where we cherry-picked code
    that had warnings from newer Rust versions (#3480).  Now that we pin
    the Rust version with our `rust-toolchain`, I don't think this is an
    issue anymore.
  - The task itself now runs `git submodule --init`

Moved the Rust installation code from `cross-compile-setup.sh` to the
new `rust.yml` task.