Removing the interrupt unittests
This is failing on CI because of issues linking to sqlite3.  I believe
it's the same issue that prevents us from having tests in support/sql.
I don't think the test is that important, so let's just remove it.  I'm
pretty sure if this type isn't Sync + Send, we'll see build failures.

Updated the rusqlite dependency to match the sql-support one.