Remove `log_query_plans` feature from `sql-support`.
The query plan format is unstable, and new versions of SQLite, including
the version now used by Firefox, is different and incompatible. When I
wrote the support for it initially, I ported it by hand it from the
visualization used in the `sqlite3` command line shell.

In practice, it helped a little. I'd be unsurprised if nobody's used it
since -- it was not really very convenient to work with. The main reason
I added it was because I didn't know how to access the functionality
from the SQLite command line shell, because of how many custom functions
we use in our queries.

Since then, I've learned that if you SQLite was built with
[`-DSQLITE_ENABLE_UNKNOWN_SQL_FUNCTION`][doc] (my default one was not,
but compiling SQLite is straightforward), SQLite will substitute custom
functions inside `EXPLAIN` and `EXPLAIN QUERY PLAN` queries (and only
inside those -- this is the whole point of that flag) with a special
"`unknown()`" function allowing the query plan to still be generated.

This is a more convenient stable way to get this functionality, as it is
amenable to tweaking the query, adding indices, etc. and seeing how the
plan changes as a result -- avoiding the need to recompile a large Rust
codebase and rerun its test suite. But more importantly, it's stable,
and does not tie you to an old version of SQLite.

[doc]: https://www.sqlite.org/compile.html#enable_unknown_sql_function