Fixes for nighly builds
Added a `route` to our nightly branch build tasks.  I think this should
mean we can view the latest branch builds via a taskcluster index.

Moved the code to calculate if we're in a nightly into the top-level
`get_decision_parameters()` function.  It seems good to put this in one
place, especially since the way we determine if we're in a nightly build
is pretty hacky.