Removed the code to add indexes to the nightly tasks
This didn't work out well in practice, since an index can only store one
task.  So if we have all the branch build task using the same index,
whichever one runs last wins.  I think what we really want is to index
the task group, but I don't think that's possible.  Probably the best we
can do is what was already there by default, which is to index the
decision task, then find the task group from that.

Also fixed the spelling of "nightly" in the test param YAML file.