More logins error updates (#5061)
Based on taking a look at Sentry I found that most reported logins
errors are not coming from the top-level error reporting, but from
`log::error()` calls inside functions.  The 2 errors accounted for the
bulk of the events:
  - Duplicate login in update
  - Sync Error: Invalid login: Login has illegal field: Origin is Malformed

Updated all the `log::error()` calls to use `report_error()` instead.

Added some more context for the 2 main errors.  For the duplicate login
one, I added if it had an origin or a realm set, since that determines
how we check for duplicates.  For the origin is malformed error, I added
a breadcrumb to log the actual `url::Parse` error variant.

The goal here is to keep incrementally improving the error reporting so
that we can eventually understand these errors.