Update the bridged_engine to not rely on sync15_traits::Payload.
This shows how I envisage sync15_traits should work for both the bridged
engine and the bso_record based (ie, sync15) engines. This patch only
touched the bridged_engine to keep the patch small with a future
patch using this shape for the rest of the crate soon.

It leans further into the "envelope" concept introduced by the
bridged_engine, but the envelope is now split from the payload
so it can be used by bso_record.

Now, instead of going to and from the sync15_traits::Payload, we just
move to and from <T> directly:

* When converting from an incoming record to a <T>, the result is
  captured in a new enum, `IncomingDeserPayload`, which allows us
  to capture incoming tombstones and invalid record. No errors are
  returned deserializing <T>, the enum variants cover all cases.

* When converting from T to an outgoing record, we will panic if
  various constraints are not met - eg, if you have specified an
  invalid GUID. All such errors are considered "static" errors
  and aren't influenced by runtime conditions, so a panic means
  we don't have to introduce new error handling for these "impossible"
  states.

It also kills the concept of "auto" fields - it's no longer necessary
to have a field called, say, 'ttl' in your record to arrange for a TTL
on the record - these are now handled by the envelope. Note that the
'id' field on the payload does remain somewhat magic.