Move the bso_record and key_bundle into the sync15_traits crate. (#5083)
(Which further reinforces that the name of that crate should be something like
`sync15-types`)

This is preparatory work for a much larger patch that will get us further down
the path to #2841 and also fix #2712. It should not change any behaviour.

As part of this, the sync15_traits errors needs to be addressed. It's not
reasonable to just create another enum that we expect clients to add to
their error types because many of the errors overlap which makes handling
them very difficult (eg, if we want to catch a serde error, you don't want
to know which crate it originated in) - so sync15 duplicates the sync15_traits
errors and can convert between them - so consumers can continue to catch
sync15::Error and still get errors which originated in sync15-traits.

sync15-traits now has a `crypto` feature - sync15 enables that feature, but
crates which just use a `bridged_engine` do not. This means that when
vendoring into Desktop Firefox, this feature isn't enabled, so we don't
pull in any of the crypto libs there.