Store the keys in the GlobalState as an EncryptedPayload instead of EncryptedBso
We stored the keys as an EncryptedBso so that the timestamp was available
"for free", but this gets in the way of moving towards different types for
incoming and outgoing server records/payloads.

We now store the EncryptedPayload and the modified timestamp explicitly.
The semantics don't change, just what concrete type we hold in memory.