Rework how BSOs work in sync15
This adopts some of the BridgedEngine model - specifically, there's now
an IncomingBso and OutgoingBso, each with a corresponding IncomingEnvelope
and OutgoingEnvelope. There's also a new formalized "content" mechanism
for serializing and deserializing payloads to and from some T, with a
sane and consistent procedure for handling tomebstones and malformed
records.

BridgedEngine and SyncEngine now work using these new types, getting us
closer to fully integrating these engines into a single trait.

This kills sync15::Payload entirely in favor of the new system.