Rename `LoginsStorageError` to `LoginsApiError`
I think this name communicates it's purpose better and is more
consistent with other components.

Renamed `LoginsError` to `Error` for the same reasons as the change from
`PlacesError` to `Error`

Changed `logins.udl` to use the fielded error style.  Updated
`LoginsApiError` to use named fields to be compatibile with this.

Removed the `RequestFailed` variant, based on the note next to it.