Reworked tabs to use the new error system (#5168)
Renamed `TabsError` to `Error` for the same reasons as the logins and
places change.

Added a `TabsApiError` enum with conversions from `Error` and
started using it in the UDL.

Removed the `TabsError::SyncResetError` variant.  That was a pretty
weird one, since it all anyhow errors were converted into it. Only
`sync/engine.rs` used `anyhow::Result`, so we could assume any anyhow
errors were sync errors.  I'm not sure why you could assume it was a
sync reset error though.  In any case, I changed it so functions in
`sync/engine.rs` use the normal `Result` enum, and we convert them to
`anyhow::Result`s only for the `SyncEngine` impl.

Removed the `TabsError::SyncAdapterError` variant when full-sync is not
enabled.  The functions that used that now return top-level `ApiResult`
values and can use `TabsApiError::SyncError`.