Be more careful with URL redaction
I thought that the initial scheme was what Desktop used, but that was
just wrong.  If we're going to be using a new method, then let's be
extremely careful about any potential PII leaks.

The initial scheme potentially leaked PII, since URLs with many path
segments would have a relatively unique redacted version.  Therefore if
you had a known URL in mind like that, you could compare the redacted
string to the redacted versions of the URLs in the Sentry breadcrumbs
and potentially determine that it was likely the user visited the URL.