Remove the not-quite-functional test-utils feature from sync15
When doing #5139 I tried to have a test-utils feature for sync15,
which didn't really work out for various reasons but it was left in.

Currently `cargo test -p sync15 --features=sync-client` fails, although
CI doesn't every run that (it runs with default and all features)

This PR kills that feature entirely.