Re-add the record ID to the BSO payload.
It turns out that in #5139 I was too agressive about removing the ID
from a BSO payload - some things in desktop got a little upset, all of
which are fixable, but this change might cause subtle breakage when
syncing to older devices.

This PR ensures that the record's ID is in the JSON payload.