Expose the `logins::EncryptorDecryptor` as an UniFFI interface
Change the logins API to use `EncryptorDecryptor` directly rather than
inputting keys.  The main benefit of this is that we only construct the
`EncryptorDecryptor` once, rather than for each API call.

Also, the autofill API is very similar to the logins one.  We could
consider moving `EncryptorDecryptor` to some shared place, like a
support crate, and using that for both.  That would avoid some code
duplication.  It also is one way to sidestep an issue we have with Swift
namespaces for a bit longer.  Currently, if we started using
`autofill.udl` on Swift we would have a name collision with
`create_key() (see uniffi-rs#1426 for details on this).