Move the places `include_scaffolding!` call to the crate root
This is what UniFfi expects and will start enforcing soon.  When I was
testing app-services with the current code, I got errors because the
`UniFfiTag` struct was not in the crate root.

Made a bunch of things `pub` that weren't before.  This seems logical,
since we're exporting those things in our UniFfi API.  It also was the
simplest way to make the import work.