Updating the nightly build code
Generalized the concept of "nightlies" as "preview builds". A preview
build can either by a nightly build from `main` or a build from a PR
branch that we want to to test against `firefox-android` /
`firefox-ios`.  Non-nightly preview builds are still a WIP.

Added method for triggering a preview build from a PR title.

Upload preview builds to the nightly maven bucket.