Don't trigger a nightly build on every push to main
Added `run-on-tasks-for` for the nightly publish task so that it doesn't
run on every push to main.  I'm not sure if this is exactly the right
configuration -- it likely will need to change again when we implement
release-promotion.  But this matches how beetmover tasks work, which
seems like a good place for now.

Added taskcluster test params for a push to main.