Update the build-summary alert slack channel
This updates the channel we send an alert to when a nightly or release
build fails.  The old value was `#app-services`, which we're not using
anymore.  I changed it to `#applicaiton-services-eng` which has a bunch
of relman/releng members in it.

The other possibility is `#sync-bots`, but I'm thinking we can reserve
that one for things like Sentry errors that happen fairly frequently.
This one should be very low volume.