Refactoring FirefoxAccount constructor
This is how our android code exports things and I think it's a nicer
API.  It doesn't allow people to specify arbitrary URLs, but I don't
think that's being used at all.

This is the start of a series of commits with unimplemented code,
intended for discussion-only.  To that end, I deleted the
`fxa_client.udl` file, since I think it would be more distracting then
useful.