Move the device initialization info to the `FirefoxAccount` constructor
It seems simpler to pass this in to the constructor rather than
requiring the app to call a method "soon after".  I'm guessing the
reason had to do async-related issues when loading the profile and I'm
hoping it can go away with async callbacks.

An unfolding pattern for the construcor is all data is passed in with
the `FxaConfig` parameter but the callback interfaces are passed in
separately.  This feels good to me, but I'd be open to other solutions.