Refactoring the auth module and adding event handler code
Consumers will now pass in an OAuthHandler that handles the flow, which
simplifies the API in several ways.

Added more complexity to the state tracking.  Instead of just tracking
if you're logged out or logged in, it also tracks transient states like
logging in or loading the profile, as well as states like
reauthentication needed.  The goal is to allow apps to update their UI by
simply listening to state-change events.

Refactored handle_push_message() to not return a result.  Instead it can
trigger any events needed.