Updating the code to handle push messages in a separate process
This is a complicated issue that happens on iOS.  On that system, we
need to use a separate process to handle the push message, which means
that we can't update any state or emit any events in the main process.

I think this API has more-or-less the right shape, but still needs some
fleshing out.