Add a debug-tools feature to sql-support.
It steals a cute function from webext-storage which can dump any
arbitrary query.

It also allows for the use of a `dbg()` sql function which can be
helpful trying to debug SQL as it is executing. The function is also
available as a no-op when the feature is not available to prevent us
accidently using it in released code.

Sadly this required updating the use of prettytable-rs as the old
version crashed on Apple silicon, which had a bit of a blow-on effect
on other dependencies.