Revert "Remove support for migrating from pre-app-services fxaccount implementations"
This reverts commit e2598e30b34e017ecc20fdba553603916546d4ef.

This commit seems like the right thing to do, but it's causing build
issues with firefox-ios
(https://app.bitrise.io/build/ca2da5d9-ffe2-4a3a-ad76-2e87c7d7cedf).  I
think the safest thing to do is revert it for the 116 branch and fix the
firefox-ios breakage in the 117 branch.