New branch build system.
Instead of trying to run all the tests, let's just try to build the
application.  This system will hopefully be simpler to maintain and also
New branch build system.

We lose a little by not running the unit test, but we keep the main
features of the branch build system: testing if there are unresolved
breaking changes and giving devs/testers an APK to manually test.

This also made it much simpler to support firefox-ios, which is now also
supported.