Adding the `StateManager` struct
This acts as a container for the FirefoxAccount's state. All fields are
private, which means any state mutations need to go through one of its
methods.  This makes it easier to follow the mutations.

As part of this refactor, I tried to make the names be more precise.
For example `last_sent_device_capabilities` instead of
`device_capabilites`.

Renamed `State` to `PersistedState` and the current "PersistedState"
enum to "PersistedStateTagged".  My hope is that these names make it
clear what each type is doing.

Removed the `RefCell` around `FxaTelemetry`.  It doesn't seem useful
when the entire struct is wrapped in a Mutex.

Removed some dead code