Reworked handle_oauth_response()
Moved all the code that updates `self.state` to the bottom of the
function.  This makes the changes atomic, which is easier to understand.
Either we successfully handle the response and update our state, or we
fail to handle the response and the state stays the same.

I don't think this will affect the functionality.  If we fail to handle
the oauth response, then we will be in a disconnected state and any of
the changes that would have previously been made won't matter. Please
double-check my reasoning here though.