Added the `StorageHandler` callback interface
This is a simpler way for consumers to handle saving the fxa-client
state.  Whenever the state needs saving, the JSON data is passed to
`save_state()`.  This removes the need for us to explicitly list each
method that changes the state in the docstrings and for consumers to
read and keep up-to-date on those docstrings.