Adding device functionality
Added DeviceList type, which is basically the `DeviceConstellation` type
that our consumer apps currently have to define in their wrapper code.
Let's update them to use this type instead and remove the need for the
wrappers.

Added a standalone `parse_push_message()`, which can be used on iOS to
handle push messages in the notification process without actually
constructing a FirefoxAccount.

Renamed `GetDeviceResponse` to `Device` in `internal/device.rs`. Having
both `Device` and `internal::device::Device` as two distinct types is
confusing.