remote_settings: Don't hold the state mutex during server info requests.
The server info request to look up the attachments base URL can take an
unknown amount of time, and holding the mutex until it's done blocks
requests for other attachments and records.

With this change, it's possible for concurrent attachment requests from
a new client to request the base URL multiple times, though we'll still
cache the URL for subsequent requests.