Make toolchain-desktop-macos depend on fetch-nss-artifact [ci full]
The `toolchain-desktop-macos` task implicitly depends on the
`fetch-nss-artifact` task, as it is using the index here:
https://github.com/mozilla/application-services/blob/41ecfc79dd1ea3c5962967707e11e0497a7e7703/libs/build-nss-desktop.sh#L61

However, it doesn't list the task as a dependency. This means that the
`fetch-nss-artifact` task never gets pulled into the graph and needs to be
run manually anytime it expires or is updated.

By adding this as a `fetch` to the `toolchain-desktop-macos`, it:

1. Sets up the dependency link in Taskcluster
2. Automatically downloads + extracts the artifact (so we can cleanup
   the build script a bunch)